Move K8Dns discovery to a separate package
